// src/lib/whatsapp/generated/message-templates.d.ts
export interface components {
    schemas: {
// ...
        WhatsAppBusinessHSMStatus: "APPROVED" | "ARCHIVED" | "DELETED" | "DISABLED" | "IN_APPEAL" | "LIMIT_EXCEEDED" | "PAUSED" | "PENDING" | "PENDING_DELETION" | "REJECTED";
// ...
        WhatsAppBusinessHSMQualityScore: "GREEN" | "RED" | "UNKNOWN" | "YELLOW";
        /**
         * @description Reason the template was rejected
         * @example NONE
         * @enum {string}
         */
        WhatsAppBusinessHSMRejectionReason: "ABUSIVE_CONTENT" | "CATEGORY_NOT_AVAILABLE" | "INCORRECT_CATEGORY" | "INVALID_FORMAT" | "NONE" | "PROMOTIONAL" | "SCAM" | "TAG_CONTENT_MISMATCH";
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}